python-projects/credit-score/pyproject.toml
[project]
name = "credit-score"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "pydantic>=2.12.5",
    "scikit-learn<1.8.0",
    "pandas<3.0.0",
    "torch>=2.0.0",
    "joblib>=1.5.3",
    "python-json-logger>=4.0.0",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "mlflow>=2.15.1",
    "seaborn>=0.12.0",
    "matplotlib>=3.7.0",
]

